refactor(statsd): wrap client in arc
