+++ libdd-data-pipeline/src/trace_exporter/builder.rs
@@ -21,7 +21,7 @@ use crate::trace_exporter::{
 use arc_swap::ArcSwap;
 use libdd_capabilities::{HttpClientCapability, LogWriterCapability, MaybeSend, SleepCapability};
 use libdd_common::{parse_uri, tag, Endpoint};
-use libdd_dogstatsd_client::new;
+use libdd_dogstatsd_client::DogStatsDClient;
 use libdd_shared_runtime::SharedRuntime;
 #[cfg(not(target_arch = "wasm32"))]
 use libdd_shared_runtime::{BlockingRuntime, ForkSafeRuntime};
@@ -583,7 +583,7 @@ impl<R: SharedRuntime> TraceExporterBuilder<R> {
 
         let dogstatsd = self
             .dogstatsd_url
-            .and_then(|u| new(Endpoint::from_slice(&u)).ok().map(Arc::new));
+            .and_then(|u| DogStatsDClient::new(Endpoint::from_slice(&u)).ok());
 
         let base_url = self.url.as_deref().unwrap_or(DEFAULT_AGENT_URL);
